Upgrade to .net 6 and fix error in trainigs command
Training had an error for the day. If you set one, you got an cast error, because they tried to cast it to nullable long from int

+++ src/DiscordBot.Commands/Modules/Slash/Trainings/TrainingsSlashModule.cs
@@ -25,7 +25,7 @@ namespace DiscordBot.Commands.Modules.Slash.Trainings
         public async Task GetTodayTraining(
             InteractionContext context,
             [Option("day", "Choose the day you want to print.")]
-            long day = -1,
+            int day = -1,
             [Option("traini", "test")] Trainingsplan plan = Trainingsplan.SaschaHuberPlan1Starter)
         {
             try
